stagefright: Configure codecs correctly in the mediatek platform
We were setting the width/height as stride/sliceheight for both
MTK and software codecs, which is wrong, it is only required in the
former and will cause visual corruption in the later.

Additionally, this adds buffer alignment which prevents MTK codecs
from following a slower path.

Finally, we're now also using stride/sliceheight whenever extracting
a video frame.

Change-Id: I3c40e17ad1757d3b27d919bf5378974b971aaacf